refactor(renderer): extract ModCardHeader from ModCard
