+++ src/renderer/src/components/mod/mod-card.tsx
@@ -1,25 +1,13 @@
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@renderer/components/ui/alert-dialog";
 import { Badge } from "@renderer/components/ui/badge";
-import { Button } from "@renderer/components/ui/button";
 import { Separator } from "@renderer/components/ui/separator";
 import { cn } from "@renderer/lib/utils";
 import type { ModInfo } from "@renderer/types/mod";
 import { formatDate, formatSize } from "@shared/utils";
 import { useRouteContext } from "@tanstack/react-router";
-import { CalendarIcon, FolderIcon, TerminalSquareIcon, TrashIcon } from "lucide-react";
+import { CalendarIcon, FolderIcon } from "lucide-react";
 import { memo, useRef } from "react";
-import { useTranslation } from "react-i18next";
 import { toast } from "sonner";
+import { ModCardHeader } from "./mod-card-header";
 import { ModIniList } from "./mod-ini-list";
 import { ModPreviewContainer } from "./mod-preview-container";
 import { getModColorClass } from "./utils";
@@ -43,7 +31,6 @@ export const ModCard = memo(function ModCard({
   onToggle,
   onToggleKeyUpdate,
 }: ModCardProps) {
-  const { t } = useTranslation();
   const { queryClient } = useRouteContext({ from: "__root__" });
   const mouseDownTargetRef = useRef<EventTarget | null>(null);
 
@@ -112,19 +99,6 @@ export const ModCard = memo(function ModCard({
     }
   };
 
-  const handleDelete = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    const promise = window.api.invoke("util:fs:trash", mod.path);
-    toast.promise(promise, {
-      loading: "휴지통으로 이동 중...",
-      success: "삭제 완료",
-      error: "삭제 실패",
-    });
-    promise.then(() => {
-      queryClient.invalidateQueries({ queryKey: ["modGroup", selectedGroupPath] });
-    });
-  };
-
   return (
     <div
       className={cn(
@@ -152,56 +126,7 @@ export const ModCard = memo(function ModCard({
         </div>
       )}
 
-      <div className="flex items-center justify-between pb-1 relative z-10">
-        <span className="text-sm truncate font-semibold">
-          {mod.name.replace(/disabled/gi, "").trim()}
-        </span>
-        <div className="flex items-center gap-1">
-          <Button
-            variant="ghost"
-            size="icon"
-            className="size-7 hover:bg-accent/20"
-            onClick={(e) => {
-              e.stopPropagation();
-              window.api.invoke("util:openCmd", mod.path);
-            }}
-          >
-            <TerminalSquareIcon />
-          </Button>
-
-          <AlertDialog>
-            <AlertDialogTrigger asChild onClick={(e) => e.stopPropagation()}>
-              <Button variant="ghost" size="icon" className="size-7 hover:bg-accent/20">
-                <TrashIcon />
-              </Button>
-            </AlertDialogTrigger>
-            <AlertDialogContent onClick={(e) => e.stopPropagation()}>
-              <AlertDialogHeader>
-                <AlertDialogTitle>{t("page.mod.dialog.delete-mod.title")}</AlertDialogTitle>
-                <AlertDialogDescription>
-                  {t("page.mod.dialog.delete-mod.description")}
-                </AlertDialogDescription>
-              </AlertDialogHeader>
-              <AlertDialogFooter>
-                <AlertDialogCancel>{t("g.cancel")}</AlertDialogCancel>
-                <AlertDialogAction onClick={handleDelete}>{t("g.delete")}</AlertDialogAction>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialog>
-
-          <Button
-            variant="ghost"
-            size="icon"
-            className="size-7 hover:bg-accent/20"
-            onClick={(e) => {
-              e.stopPropagation();
-              window.api.invoke("util:openPath", mod.path);
-            }}
-          >
-            <FolderIcon />
-          </Button>
-        </div>
-      </div>
+      <ModCardHeader mod={mod} selectedGroupPath={selectedGroupPath} />
 
       <div className="flex flex-row h-[calc(100%-2rem)] space-x-2 relative z-10">
         <ModPreviewContainer
